move sdk package to go/sdk
