+++ examples/language-analyzer.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc"
 	"gopkg.in/bblfsh/client-go.v2/tools"
 	log "gopkg.in/src-d/go-log.v1"
-	sdk "gopkg.in/src-d/lookout-sdk.v0/go"
+	"gopkg.in/src-d/lookout-sdk.v0/go/sdk"
 	"gopkg.in/src-d/lookout-sdk.v0/pb"
 )
